Exploit version of a single telnet auth module
Like with our SSH known public key modules, we have a need for a single
known user:pass cred for telnet. A couple things pop up with this sort
of vulnerability.

Sadly, I haven't figured out how to register the session opened
correctly, and I'd prefer not to haul over a million extra lines of code
for cred management and scan management when I'm just trying the one
user:pass against the one target.

So, the vuln is recorded, everything seems to work, and then the session
dies.